chore: config: split flags processing
Split the command line flags processing from the
orchestrating flow in its own function.
Trivial code movement and rote conversions.
